refactor: update name & webpack watch poll config
