+++ webpack.common.js
@@ -164,5 +164,7 @@ module.exports = {
   },
   watchOptions: {
     ignored: /node_modules/,
+    poll: 1000,
+    aggregateTimeout: 600,
   },
 };
